VisionSeal-Refactored/requirements.txt
# VisionSeal Complete - Refactored Dependencies
# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-magic==0.4.27

# Database & Storage
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# AI & Vector Storage
openai==1.3.7
weaviate-client==3.25.3
tiktoken==0.5.2

# Document Processing
python-docx==1.1.0
PyPDF2==3.0.1
python-pptx==0.6.23
openpyxl==3.1.2

# Data Processing
pandas==2.1.4
numpy==1.24.4

# Web Scraping (for automation engines)
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
pathlib==1.0.1
aiofiles==23.2.1

# Logging & Monitoring
structlog==23.2.0
colorama==0.4.6

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.0

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Environment Management
python-dotenv==1.0.0